requirements.txt
Django==2.2.5
django-bootstrap4==1.0.1
django-debug-toolbar==2.1
django-filter==2.2.0
pandas==0.25.3
tablib==0.14.0
